requirements.txt
google-cloud-bigquery==3.9.0
Flask==2.0.3
schedule==0.6.0
boto3==1.26.11
numpy==1.24.3
botocore==1.29.11
pandas==2.0.1
redis==4.3.4
gunicorn==20.0.4
newrelic==6.0.1.155
argparse==1.4.0
joblib==1.2.0
sagemaker==2.116.0
requests==2.28.1
pyyaml==6.0
gevent==22.10.2
greenlet==2.0.2